[fips] configure_fips() doesn't have a validation mechanism if bootdev isn't a UUID.
Bug#79

Current code look if a separate /boot partition exist in FSTAB, if nothing is return (nonzero value) it skip the bootdev portion but what if the $bootdev is nonzero and doesn't contain a UUID ?

The first field of /boot line could be a device name in the form of (/dev/sdX,...) or a label (LABEL=Boot).
I think iff UUID is needed, then we should add a validation mechanism.

Signed-off-by: Eric Desrochers eric.desrochers@canonical.com