token: use entitlement-specific token directive if present.
Fallback to machine-token.machineSecret if specific entitlement
does not provide a unique token for that entitlement.

This paves the way for the use of a global token to drive all
entitlements. But it may take a while for livepatch, fips and esm
backends to support such a global token.

In the iterim, allow the entitlement response from Contract API
to provide a specific token directive which have be used if the
global token isn't plumbed yet.