tox: use nose2, speed up run, and add coverage
This utilizes nose2 so we can also turn on coverage for the python
files. By setting the envdir the tests also run considerably faster.
The file was cleaned up to make it easier to read.

The setup.py was added to tox and corrections made to clean up lint errors.

The *requirements.txt files were alphabetized and cleaned up.