Enable Trusty build, remove python-six, fix flake8
Trusty is now buildable in an sbuild.

This removes dependencies on python-six, which was getting used
throughout. This is meant to be a Python 3 only application so there is
no need for python-six. Here is how various usage was updated:

python-six     --> python-3
string_types   --> str
binary_types   --> bytes
StringIO       --> io.StringIO
six.moves      --> import urllib or input()
@add_metaclass --> function signature

Due to differences in the way that flake8 can be installed across Trusty
to Bionic, this simplifies the dependency list and deals with the
different mechanisms for calling it. The trusty found flake8 errors are
fixed and one error is ignored.

more