add flake8-{trusty,xenial,bionic} tox targets
These use a constraints file per-release to install the flake8 and
pep8/pycodestyle versions that were present in that release.

This change also makes it possible to run something like `tox -e
py3-trusty` which will use the trusty constraints file for the regular
tests; for this to be useful, we would have to expand the packages
listed in the constraints file to include all of our dependencies.

(This generative syntax was introduced in tox 1.8, which is present in
xenial and later.)