tests: handle versions of pytest without caplog fixture
The versions of pytest in both trusty and xenial predate the caplog
fixture's introduction.  In xenial, we could use the pytest-catchlog
package (which is the code that was integrated in to pytest itself
between xenial and bionic), but that package isn't available in trusty
so we still need a shim.