keyrings: remove uncessary and only install ESM
Keyrings no longer need to be shipped with the package, therefore we
should remove them from the source code of the UA client. The exception
however, is for ESM, in which case we need to ship it, install it, and
remove it during a purge.

The fingerprint of the included ESM key is:
74AE 092F 7629 ACDF 4FB1  7310 B4C2 AF7A 67C7 A026

Fixes #256
Fixes #334