Clean up dev directory
This breaks the dev directory into 'tools' and 'demo'. Where the tools
directory has building and test related files and the demo directory has
files related to demoing and development of the package itself.

Python files in tools were all made pylint clean and verified to be
working correctly.

This also updates the related Makefile, tox.ini, and README.md files
with the changes in directories.

Fixes #255