rules: override dh_auto_clean
dh was told to use the pybuild buildsystem, and that means that in the
dh_auto_clean step it will call setup.py, and not the Makefile. Turns
out the setup.py clean action doesn't do all that's needed here, so we
are better off overriding this and using the Makefile's clean target
which is more complete.