repo: persist a localEnabled flag in machine-acces for non-root status
Any repo-based entitlement can not check apt-cache policy to determine
if the entitlement is enabled because apt will automatically pin -32768
for a non-root user of an authenticated apt source because the non-root
user does not have permission on /etc/apt/auth.conf.d/<secureconffile>
due to mode 600 on those files.

As a result uaclient will fallback to our localEnabled flag in public
cache files which is set True after enable and False at disable.

Fixes: #339