Drop ubuntu/ from hostname bit in auth.conf
Trusty is picky regarding the format of the lines in auth.conf,
see https://github.com/CanonicalLtd/ubuntu-advantage-client/issues/275
for details.

We need one entry with just the hostname, and a trailing slash ("/").
If there are more entries to be added, the more specific ones need to be
at the top of the file. Trusty, at the moment, has no other repository
backed UA services, so this particular bit isn't of concern (FIPS is only
available in xenial and later).