test_apt: don't pass new to mock.patch
On trusty, I've observed a bad interaction between invoking mock.patch
with a new object and pytest fixtures.  (Invoking mock.patch with a new
object means that it doesn't pass its object in to the decorated
function; this can make test code a little cleaner, but this code path
is the one that causes exceptions).

This commit modifies the mock.patch decorators on test methods using the
remove_auth_apt_repo_kwargs so that their (unused) mocks are passed in
to the test function nonetheless.