tools/constraints-trusty.txt: revert to trusty versions
We had massaged the constraints file to use versions that would work on
releases later than trusty (in particular, bionic) for developers.
Unfortunately, this meant that Travis wasn't testing with trusty
versions, meaning we weren't catching potential regressions.

This reverts us back to the versions in trusty; we'll need to consider
other methods for developer machine testing of older releases.

Fixes #398.