logging: set default to string
The default value for logging when it was missing was becoming an
integer (logging.INFO == 20) and then being passed around between
functions that expected a string value for the level.

Rather than storing logging.INFO as the default value for logging, use
the same value/string that we print to the file, namely 'INFO' and then
do the right thing with the string.

Fixes #406