attach: suppress messages from inapplicable default entitlements
`ua enable` will emit messages when determining if an entitlement is
applicable to the current system.  For a less noisy user experience, we
don't want to emit these messages when enabling default entitlements
during `ua attach`.  This patch silences that output (but retains the
output from entitlements that are applicable).

Fixes #411.