auto-enable: move enableByDefault handling up into UAEntitlement
Shift repo-based enableByDefault handling into parent class so
livepatch can share this functionality.

Add allow_enable boolean flag to process_contract_deltas which
defaults to False. When True, auto-enable an inactive service if the
delta contains a transition to enableByDefault: True. When allow_enable
is False, print a message to the console and logs that a recommended
default entitlement is not auto-enabled

allow_enable will only be True from `ua attach` operation.