uaclient: introduce separate concept of application_status
Instead of having a single operational_status for each entitlement, this
commit introduces the concept of a separate application_status; this
represents whether or not the entitlement in question is currently
applied to the system (independent of whether or not the machine is
entitled to have the entitlement applied to the system).

With the introduction of UAEntitlement.application_status as an abstract
property, all of the entitlement-specific logic was moved out of
the operational_status implementations on UAEntitlement subclasses, so
it was moved to be a concrete method on the superclass.