uaclient: introduce ApplicationStatus.PENDING
This is a third ApplicationStatus that represents "enabled but not yet
fully operational".  This commit updates the code paths that deal with
ApplicationStatus to treat ApplicationStatus.PENDING as equivalent to
ApplicationStatus.ENABLED; we don't yet have any code paths that diverge
based on it.

(The first examples of this will be FIPS/FIPS Updates; they aren't fully
enabled until you have rebooted into the linux-fips kernel.)