subp: reduce apt-helper call timeout from 2 mins to 20 seconds
apt has a default timeout of 2 minutes for tools like apt-helper.

Add a timeout parameter to util.subp allowing ua client to handle
subprocess.TimoutExpired errors for specific commands.

Apply a 20 second timeout to the apt-helper call used when validating
credentials against an apt repo url. If we don't see a response from the
repo in that apt-helper call in 20 seconds, chances are we something is
wrong that a 2 minute timeout isn't going to resolve.

Fixes: #626