contract: apply_series_overrides before computing deltas
Our orig_access has already had the series overrides applied, which
means it will never be equal to the contract server's response.  As a
result, we would _always_ determine that there was a delta, and apply
it.

This is simply fixed by applying the series overrides to the new access
dict before we determine if there are any deltas.

Fixes #666