gpg: do not create .gnupg dir during key extraction from shared keyring
Both debian/postinst and uaclient/gpg call to gpg on the command line in
order to extract a known GPG key for a service we are enabling.

Update both of these callsites to use the same gpg parameters.
Also, add --no-options parameter which prevents custom gpg config
options from being sourced as well as preventing the creation of
.gnupg in the current user's homedir.

Fixes: #776