livepatch: check is_container before recommending kernel upgrade
On a container, livepatch isn't every going to be supported, so
there is little sense in checking kernel version first and telling
the user to upgrade their kernel only to find out after reboot that
they still can't enable livepatch because livepatch doesn't run on the
container.

Reorder static_affordances (which tend to be uncorrectable on a
platform) before more dynamic environment affordances like <series>
or kernel revision which can up updated on a machine to bring it into
compliance.

Addresses part of #807