status: custom descriptions for n/a service status
The response from machine attach (machine-token.json) provides an
availableResources list.  Any item in availableResources which is not
"available" will provide a "description" that should be surfaced as a
user-facing custom description for any n/a service.

The command `ua status` should surface these custom descriptions in the
DESCRIPTION column for any n/a service.

In the absence of a custom status "description" for an inapplicable
service, use the static entitlement.description.

Fixes: #790