d/postinst: remove client upgrade logic
The client does not need to deal with upgrades for Trusty. The previous
version of the client never knew about a cache and therefore this logic
is working around a future issue that may or may not exist.

This logic can be re-introduced we a client for Xenial, Bionic, or other
releases is worked on.

Fixes #881