entitlements: rely solely on contract server for repo_url
Before this commit, we had a hard-coded default that we would use if the
contract server failed to provide us with an aptURL directive.  The
logic for that default becomes more complex with xenial/bionic, because
the URLs have changed.

Instead of introducing more complexity into the client, we instead drop
that defaulting, with the assumption that the contract server will
always provide an aptURL directive.