esm: don't disable_apt_auth_only for ESM entitlements
On trusty, we want to advertise the availability of ESM Infra updates.
We don't want to do that for xenial and later, so we don't need to set
disable_apt_auth_only on ESM entitlements.

Note that this change will break behaviour on trusty, so will need to be
elaborated (per the inline TODO) before it will work on all supported
releases.