repo: match strict repo url in apt-policy to avoid esm substring matches
ESM-Infra aptURL directive is currently a substring of ESM-Apps.

The application status for a repository-based service, such as
fips*, esm* and cc-eal is determined by searching for an active aptURL
directive for that service matched in the apt-policy on the machine.

Make that url match more strict so only the full and specific
<aptURL>/ubuntu is matched to determine whether the service is ACTIVE.