pro: rename premium -> pro and attach-premium -> auto-attach
Ubuntu Pro is the marketing name for the images that are entitled to
multiple Ubuntu Advantage support services. The action that ua client
performs is auto-attach of those entitled services.

Rename the CLI command attach-premium to auto-attach.
Update user-facing messages replacing Premium with Ubuntu Pro
Update functions and methods to avoid premium naming.