d/rules: don't install the apt hook
The hook only reports for esm-infra, not for esm-apps.  We don't have
time to address that before the deadline we are working towards, so drop
it to avoid an inconsistent user experience.

(Note that for xenial and later, the hook only reports when entitlements
are enabled, so this will retain the unattached/un-enabled behaviour we
already have.)