attach: persist any status config changes after attach failures
There can be cases where auto-enablement of a service results in an
error. Or, intermittent connectivity to ua-contracts routes end up
resulting in an error during `ua attach`.

In the event that attach succeeds, but auto-enablement actions fail
status needs to be persisted to reflect whether attach succeeded and
any services have been successfully enabled.

Fixes: #952