pro: add postinst logic emit debconf note to users during pro upgrade
There are package breaking changes due to migrating init scripts out
of ubuntu-advantage-tools and into ubuntu-advantage-pro.

In the scenario where an existing Azure Ubuntu Pro instance upgrades
from 19.7 to 20.2, the system will only upgrade the existing ua-tools
package and it will be left without knowledge of ubuntu-advantage-pro.

This debconf note in postinst checks for this condition and alerts the
user to a manual post installation step that needs to take place.

Without ubuntu-advantage-pro package, systems will not be able to
auto-attach across reboot if they were to become detached due to a
`ua detach` command.