entitlements: static_affordances runtime instance property
Entitlement.static_affordances need to be a runtime instance property
so that FIPS* can interactively prompt the user with a callable
prior to performing an enable or disable per #1031.

Once the support is there for dynamically calling external functions
FIPS can use util.prompt_for_confirmation before enable or disable
of the service.