Use additionalPackages from service payload (#1064)
Right now, ua-client is not using the additionalPackages field when enabling repo services, like cc-eal and fips. This PR adds the possibility for ua-client to start using these data when enabling repo services.

For fips, we extend packages()  to understand which conditional extra packages should be installed when enabling the service.