Disable beta services to be show/enabled without flag (#1069)
Right now, services that are in a beta stage, such as fips are shown normally when an user run ua status. We are now changing that behavior for ua status to only show production ready services. If the user wants to see all services, the command to run will be ua status --beta.

Current beta services are: cc-eal, cis-audit, fips and fips-updates

Also, if the user tries to enable a beta service using: ua enable fips, ua-client will raise an error saying that the service could not be found. To enable beta services, users will need to run the following command: ua enable fips --beta

Fixes: #1079