Update disable and enable commands (#1055)
This PR creates an initial abstraction to allow multiple entitlements to be enables/disabled using the uaclient command. Right now, the idea is to try to enable/disable as much entitlements as we can. This means that we will not abort if we find an invalid entitlement or cannot disable/enable one of them.

To be more clear, this is an example of the output of the following command:

`ua disable ent1 ent2 ent3 --asume-yes true`

Where:

* ent1 is an invalid entitlement
* ent2 is not enabled
* ent3 is enabled

The command output will be the following:
```
ent2 is not currently enabled
See: sudo ua status
Updating packages lists
Cannot disable 'ent1'
For a list of services see: sudo ua status
```