tests: create /tmp/pr_source.tgz when not in travis and build_pr=1
When specifying UACLIENT_BEHAVE_BUILD_PR=1 in developer environment
the behave tests expect to find a compressed copy of the current
source tree at /tmp/pr_source.tgz because .travis.yml does this.

If running locally, and no TRAVIS environment variable is present,
tar the current working dir to /tmp/pr_source.tgz so the integration
tests will build a deb in and use it.