Override ua config for beta services (#1100)
As stated in #1080, we need a way to override some configurations for uaclient. This is an initial step in that direction, by allowing users to override the beta behavior in uaclient. This PR allows for users to visualize all beta service by adding the following lines to uaclient.conf:

features:
  allow_beta: true