Do not enable services if blocking services is active (#1113)
Users should not be able to enable livepatch if fips is active or fips if livepatch is active as stated in #1029 . We are now add the possibility for verifying if blocking services are enabled before activating a service. If this happens, we will not enable the service and show the user which services are responsible for that behavior.

Furthermore, we are only testing this feature now in Bionic. Although there are some improvements we can achieve for testing that feature, there are some potential issues that I have found (#1111 and #1112) that are posing a difficult for developing vm tests

Fixes: #1029