Improve pro testing for esm package installs and apt policy (#1142)
In the pro images test, verify if the both esm-infra and esm-apps are properly configured and packages originated from those sources can be downloaded and installed without an issue.

This is a working solution, but there is still a discussion about if we should add dummy packages to both of these origins in an attempt to simplify and guarantee a more reliable test for pro images. But since we have not reached a consensus on it yes, this is a working solution nonetheless, although it suffers from some flaws, like checking the package version directly for esm, which is not mandatory for the package maintainer to add when uploading a new package.