Fix enable fail bug
Currently, if a service fails when being enabled, we will
show an error messaging stating that the service could not
be found. That is because we treating beta service exceptions
exactly the same as we treat other enable errors. We are now fixing
that behaviour by differentiating beta errors from other
exceptions that may happen when enabling a service.