Save machine-token when user runs detach
Now when a user detach from a subscription we will not disable any of
the enabled services and we will create a copy of the current machine
token file on /var/lib/ubuntu-advantage/private/machine-token.json.past

When a user want to attach to a new subscription, we will compute
the contract delta based on the past machine token and the current one.
We will use that to disable services that the user is no longer
entitled.