Show error when enabling service with invalid repo
When we enable a service and a user has an invalid repository in the
system, apt update commands will fail to run. Because we run
apt update when we enable repo type services, we stumble upon this
issue and fail to enable the service. However, we are not showing
the user the reason why the enable command failed. We are now updating
the logic to display this apt issue when this problem happen during an
enable operation.

Fixes #954