Add upgrade scenario tests when FIPS is enabled
Add FIPS tests for xenial and upgrade test from xenial
to bionic when FIPS is enabled in the system. However,
FIPS Updates is not currently working when upgrading
from xenial to bionic due to a problem on do-release-upgrade
where we fail to access the right repo for FIPS Updates
during the upgrade and end up disabling the ppa, causing
some package, e.g. ubuntu-fips, to be removed.