Add Google Cloud Platform support (#1269)
* Add Google Cloud Platform (GCP) support.

This adds the ability to detect if on a Google Cloud Platform system
using a combination of the cloud-init datasource and DMI data using the
product name. The client will retreive a JWT token and send it to the
contract service for verification.

A note on naming, the contract service is expect 'gcp'. cloud-id and
cloud-init datasources use 'gce'.

* Update uaclient/clouds/identity.py

Co-authored-by: Chad Smith <chad.smith@canonical.com>

* mypy: fix absract class errors

Co-authored-by: Chad Smith <chad.smith@canonical.com>