Fix invalid ppa integration test
We have a test to verify if uaclient is behaving
as expected if we have an invalid ppa in the
system. To produce that invalid ppa, we add the
uaclient daily ppa and we use a sed to edit it.
However, the current ppa we are using is generating
a commented source file. We are now updating the
ppa we use to fix that.