Fix upgrade scenario tests for fips
Some commands were failing on the fips service upgrade
integration tests. The first one was related to using
a wildcard in the egrep command, which is not supported
on the current way we shellify commands on pycloudlib.
Also, we are now covering possibilities were we need to
downgrade a package on dist-upgrade or that a package
requires reboot after upgrading it. Additionally,
we are now dropping the past fips-updates upgrade test,
since now the fips metapckage is no longer being
uninstalled after a do-release-upgrade for fips-updates