Add pytest constraint for mypy tests
pytest was recently updated to a version that breaks the mypy tests. The
reason for that is because the new version of pytest uses a type
annotation that is only supported on python3.6 and later. Therefore
we cannot run all of the mypy tests using the latest version. Because
of that we are now adding a pytest constraint for mypy tests