Fix trusty check on postinst
On postinst we were using a grep command to
verify if the previous pkg version was a trusty one.
However, that grep command was making the postinst
script to exit non 0 in case we were installing the
package on non-trusty releases. We have now updated
the grep logic to avoid that problem