Fix log when setting apt files on repo entitlement
Right now, we have a mechanism to verify if we should apply or
not an apt delta url. If we verify that the apt url is already
present in the apt file, we do not perform any setup, since we
will understand that the setup was already performed. However,
regardless of that check, we are always logging that we will
perform a apt setup, which is misleading. We are now fixing that.