Do not install pro debs on non-pro instances
Currently, when performing integration tests, we install
both the default package and the pro package in the same
instance. However, now that we are running the
auto-attach init script before cloud-init process
the userdata, we are no longer ensuring that this script will
not run. Since the auto-attach command will fail on
non-pro instances, we must avoid installing the pro
package on those instances.