Disable azuregeneric test from Jenkins
Currently, we are experiencing issues when we try
to ssh into an Azure instance. Says this is
a basic step of the azure tests they are all
failing. To not block unrelated work to land,
we are removing that job from Jenkinsfile
until we can figure out what is happening with it