Pin importlib-metadata for mypy tests
The lib importlib-metadata was recently updated bringing
a typing annotation that is not supported by python 3.4.
This is problematic because our mypy tests covers this python
version. To allow the mypy tests to run as expected, we are
now pinning the importlib-metadata to a version that does
not break the mypy tests.