Allow uaclient to disable services during enable
When we are enabling a service in uaclient, we may
end up finding that we have an imcompatible service
already enabled in the system. Currently, we alert
the user about this issue and abort the enable operation.
We are updating that to allow the user to decide if
the incompatible services should be disabled or not
before enabling the requested service