Print error message when behave run command fails
When running the behave tests in Jenkins, it is pretty hard
to debug why one of the steps failed. The reason is that we only
output that the command returned a status different than 0. Now
we are also going to print the stdout and stderr returned by the
command in that situation, allowing us to better debug this issue