Block enabling FIPS on clouds using Xenial
Currently, Azure and GCP do not have fips specific kernel versions
for the xenial kernel. Because of that, when enabling fips on
those instances, we would end up installing the stock fips kernel
on them, which is problematic, since that kernel is not optimized
for those clouds and could generate problems for the user. Therefore,
we are now blocking enabling FIPS on Azure and GCP cloud instances