Override fips metapackage when on bionic cloud
Currently, the contract backend will always
return the ubuntu-fips metapackage as the
package to be installed when we enable fips.
However, we have specific metapackages for Bionic
Azure and AWS instances, called ubuntu-aws-fips and
ubuntu-azure-fips. Until we do not support that on
the contracts backend, we will be performing this
override directly on the uaclient code