Update metapackage override logic on fips
We are now checking the list of packages for the specific
ubuntu-fips package and this is the only package that we
will override. This means that if the additionalPackages directive
has other packages than ubuntu-fips, we still return them
in the new packages list after the override. We are also
creating a config that allows to disable this package
override