Remove expiry check from contract update (#1336)
The contract backend will now allow contract token that
have expired to keep working for 14 days. Also, the contract
backend already covers the situation were the contract is no
longer valid, which will return a 403 error that is already
treated in the code. Because of that, we are dropping the
expiry check for contracts

Fixes: #1335