Block enabling fips on GCP instances (#1343)
Currently, we do not have a GCP-optimized FIPS kernel. Installing
 the default FIPS kernel on those instances could cause
performance or even system issues. Because of that, we are
now blocking enabling fips on those instances until we have the
optimized kernel for GCP.

We can override this behavior but adding the following directive
 into uaclient.conf:

allow_default_fips_metapackage_on_gcp: true

Fixes: #1317