Remove proc file check after disabling fips
Since the new fips metapackage is now removing all
of the grub files it generated on remove, we can have
situations where after disabling fips, the user may reboot
on a non-fips kernel. If that happens, the system will not
have the /proc/sys/crypto/fips_enable file. Because of that,
we will no longer check if that file exists after disabling fips