Update how we check for fips enabled
Currently, we to check if fips is enabled, we verify if the
fips apt sources are present in the system and the fips kernel
is being used. However, if the /proc/sys/crypto/fips_enabled file
doesn't exist or is set to 0, the system will not be on fips mode.
Because of that, we are checking the proc file instead of the
running kernel to dictate if fips is enabled or not