Block detach call to contract if machine-id change
When we running a detach operation we also call
the contracts backend to delete the machine from there
as well. However, if we detect a machine id change in the system,
that call will not succeed, failing the entire operation. To avoid
that, we are now checking if the machine id in the cache and current
machine id match before doing the detach call to the contracts
backend