integration: block auto-attach on PRO base images (#1362)
Currently, we only create the disable_auto_attach entry
on uaclient.conf when we are launching the test image. This
is not working on trusty because cloud-init is editing that
file after we have already run the auto-attach script. Therefore,
we are moving the creation of this directive to the base image
(the one that will be snapshoted), which will guarantee that the
directive will be there once the test image runs