Fix bionic vm test when enable fips with livepatch
We have a bionic vm test where we enable livepatch
and we try to enable fips afterwards. On that scenario,
we also create a uaclient.conf directive to block
disabling services on enable operations. We want
to verify if in that situation, fips is not enabled
becaused livepatch is. However, we are checking the error
message on stdout, but with the new incompatible services
abstraction, that message will be on stderr if we have the
uaclient.conf directive