Add retry when running canonical-livepatch status
When running canonical-livepatch status, we may stumble on some
flaky errors. If this command fails, we will shown livepatch to
be disabled, which may not be true if it is indeed a flaky error.
To try to avoid that, we are now adding retries to that command when
it fails

Fix #1360