Update fix test on trusty machine
Currently, the fix test we perform on a trusty
machine was using a kerberos CVE to test it. However,
it seems that new trusty images are now affected by this
CVE, which is not what the test was intended to test. To fix
that, we are switching for a MariaDB CVE to use it in the
trusty test