Handle other package status when running ua fix
Currently, when we are fixing a USN or CVE, we are just handling
the case where a package has a released status. However, a package
can have different status. If we have an affected package which status
is not released, that CVE/USN cannot be completely fixed. We are
now updating the ua fix logic to handle those status and correctly
report to the user that a CVE/USN was not fixed when we found one
of the other status.

Fixes: #1435