Update output for released packages in ua fix
Currently, we list every released package in the output
of ua fix and only after that we start installing them. This
is not optimal since depending on the source of the package
we need to perform different actions to install those packages.
Therefore, we are now grouping the released packages by pocket
origin and acting on them as soon as we can.

Fixes: #1438