Check if service is enabled during ua fix
Currently, when we are running ua fix on a attached system,
we do not verify if the system has the necessary services
enabled for the fix. This is problem, since in that situation
we will incorrectly say to the user that we are installing the
package fix, which is not true. We are now checking if
the necessary service is enabled before continuing with the
package fix.

Fixes: #1462