Prompt to enable service during ua fix
Currently, when we are running ua fix and a necessary ua
service is not enabled, we abort the operation and report
that the CVE/USN was not solved. However, if the user is
entitled to that service through uaclient, we can prompt
the user to enable that service during the fix operation.
We are now adding this functionality to the ua fix command.

Fixes: #1455