fix: filter usns by cve details
When we are querying for related USNs to fix, we use find all related
USNs to a given CVE. However, the query that we are using to get the
related USNs is using applying the CVE match as a substring instead
of an exact match. This means that we will be return USNs that are not
related to our query CVE, but have are fixing a CVE which id is a
substring of our original CVE. We are filtering the returned USNs
to only those that actually contain the exact CVE id in them.

Fixes: #1470