fix: prompt for new token when expired one is used
When running ua fix we may need to install packages that
depend on a UA service. Even if the user is already attached
to a UA subscription, this may not be enough, since the token
used may be expired. If we encounter that situation, we would
need to ask the user to get a new token before proceeding with
the package fixes. We are now adding the necessary code to allow
us to handle that scenario.

Fixes: #1475