fix: show packages that were not fixed
When running ua fix when may end up in a state where
we cannot fix all the packages. If we get into that situation,
we will not display to the user what were the packages that were
not fixed and are still affected by the CVE/USN that the command
tried to fix.

Fixed: #1448