tests: run status wait before detach in PRO tests
When we are creating a PRO instance, we install the latest
uaclient package on the machine and then snapshot it to be used
as a base image for all other tests. This base image should be
a clean image, meaning that it should not be attached to uaclient.
To ensure that, we run a detach operation on the base image and config
it to block auto-attach. However, when we run detach, the auto-attach
operation may still be running. We are now running ua status --wait
before detach to avoid that.