config: create settings_overrides on config
Currently, we need cloud-init to identify if FIPS can
be correctly installed in an instance. This behavior is making
it harder to use uaclient in other scenarios, for example, when
building an image using chroot. To better handle that, we are
creating a settings_overrides dict in ua config. Users will be able
to override specific setting by just updating the uaclient.conf
file. To use these overrides, we have created a decorator that can
modify a function return based on the override defined in the config.
As of today, this is only being used in the get_cloud_type function.

Fixes: #1507