fix: update how the expired token logic works
We are now updating the logic to verify if an UA subscription has
expired. Instead of requesting an ua refresh and verifying if the
contract backend has raised any errors, we are now going directly
to the machine-token.json file and verifing until which date the
contract is valid through the effectiveTo field.