lib: fix notice handling on reboot script
When the reboot script fails, we keep the notice
that show that it is running. This happens because
the assert_lock_file decorator cannot complete and
remove the notice. We are now fixing that by catching
the exceptions in the reboot script. Furthermore, we also
issue a notice if the reboot script fails for some reason.

Fixes: #1518