d/postinst: xenial enable unauthenticated apt source for apps/infra
Refactor a couple of functions in ubuntu-advantage-tools.postinst
to a install_esm_apt_key_and_source which will reuse common logic
to emit unauthenticated APT sources.list.d files for both esm-apps
and esm-infra.

One APT sources are on disk, emit a APT preferences file that pins
the repo origins as Pin-Priority: never which tells apt to list
those packages as upgradeable, but not attempt to install or upgrade
during typical apt-get upgrade|dist-upgrade.

Ran shellcheck on this content and fixed up the suggestions
thoughout the file.

Fixes: #1529