fix: do not prompt user for beta services
When running ua fix, we may need a package fix that can only
be installed using a beta service. In that situation, we will
only permit that the fix is installed if the service is already
enabled. This means that we will not prompt user for enabling a
beta service or even attaching to UA if we found out that the
service that will provide the fixes is a beta one.

Fixes: #1544