apt-hook: print static messages on apt upgrade/dist-upgrade (#1540)
apt-hook: process template files and print messages with esm stats

apt-esm-hook will now process a set of .tmpl files in the ua messages
directory, both when apt calls it and when called separately with the
"process-templates" argument. It replaces variables in the tmpl files
with the values of ESM Apps/Infra package names and count. It then
writes the result of the replacement back to the ua messages
directory.

When called on Post-Invoke for upgrade/dist-upgrade, the hook will
print out the contents of message files if they are present.

Fixes: #1546