apt-hook: run new json hook tests on package build
The new json hook is written in golang and has some
go unit tests. We are now running these unit tests
when building a deb to help prevent regressions.

While the go program itself does not require a
particular newer version of go installed, the tests
do. The `t.Run` function we are using requires at
least version 1.7 to run. Because of this, we are
adjusting our Build-Depends appropriately. We will
now use `golang-1.10` if available (as it is in older
releases where the default version is less than 1.7),
and just use `golang` if not (in newer releases
where the default version is greater than 1.7).

Fixes: #1567