status: do not show info if not on contract
When running status on an attached machine, we present the
account and subscription info in the the end. However, if the
contract does not have that info, we will present them as empty.
We are now updating the status to not show that info at the end
if it is not present in the contract.

Fixes: #1592