Do not show beta services in autocomplete
When running ua enable, we don't want to show
beta services by default in when users hit tab and
the command autocompletes. We are now updating
the logic to only show non-beta services in that situation.
We will only show beta services in autocomplete if the
user has a uaclient.conf configuration that allow beta
services.

Fixes: #1594