logging: remove tracebacks for UserFacingErrors
UserFacingErrors are used for known situations that
our code can handle and present the correct message to
the user. They are not truly "exceptional" and do not
warrant printing a traceback when they occur. This commit
changes our logging of these types of errors to continue
to use the ERROR logging level but no longer include
tracebacks.

Fixes: #1586